position algorithm & route

+++ LOUV.Torp.BaseType/ConfType.cs
@@ -202,9 +202,9 @@ namespace LOUV.Torp.BaseType
         }
 
         public DateTime UTCTime{ get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
-        public double Depth { get; set; }
+        public float Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Depth { get; set; }
         public Target(string name="目标")
         {
             Name = name;
